items function properly
