+++ RPG Mania/Assets/Scripts/Managers/Battle/OldBattleUIManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 044f8d378687ebd42acbce1ac655ed96
+guid: b59d73983e37f90419de6b64fa47dbf8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
