+++ RPG Mania/ProjectSettings/EditorBuildSettings.asset
@@ -15,6 +15,9 @@ EditorBuildSettings:
     path: Assets/Scenes/Playable/Throne Room.unity
     guid: 9e712eca0d89a4b46bd7bbc315501996
   - enabled: 1
+    path: Assets/Scenes/Cutscene/Throne Cutscene.unity
+    guid: 89f472ce306210847b7059f0cca92b85
+  - enabled: 1
     path: Assets/Scenes/Playable/Wilds.unity
     guid: 0ce912e1b4d6e8e4880f24bdbd0ce4cc
   - enabled: 1
